Really fix marker generation
Fix #3 for real this time, I hope.

The tracing logic is revamped to include the stub None root, to generalize
the metadata calculation process. Metasets of root requirements are no
longer populated specially, but derived from the stub root's empty
metaset.

This affects quite a few things in the implementation, but those changes
are generally trivial. For exchange, I can finally correctly reason with
the metadata generation process, and make the bug disappear (I hope).

The real offender is in _build_metasets(). The line that gets the
requires_python value::

    python = pythons[parent]

Should have used the current key instead, i.e.::

    python = pythons[key]

Also re-locked with this new implementation.