Revise source dir logic
Do not save the source if the requirement is not editable -- create
a temporary directory to store the source for build instead.

For editables, if we're not in a virtual environment, and the user
does not specify a source directory, use the current directory to
store the editable requirement. This matches pip's behavior (the
previous behavior was to use a temporary directory, and throw it
away after installation).